Use the Nate Coraor standard Galaxy technology stack.
For more information on deploying Galaxy using this stack see - https://wiki.galaxyproject.org/Events/GCC2014/TrainingDay/AdminWalkthrough.

Replace Apache with uWSGI + nginx (this has advantage of using multiple processes really easily and should be a little more performant). Likewise, instead of using OS services - this starts services with supervisor - this is manages all the services which simplifies startup.sh and provides and will restart processes that die for instance. Finally - this replaces Debian Wheezy with Ubuntu 14.04 - the operating system targetted by Galaxy training material and CloudMan (and for which Nate builds .deb packages).

I think there is value in usegalaxy.org running the same technology stack that CloudMan, Docker images, and training materials use. It should simplify support and it is a stack that the devteam is confident can scale well.